cleanup migration and add old chart
