+++ App/Views/OverviewView.swift
@@ -16,6 +16,8 @@ struct OverviewView: View {
                     
                     if #available(iOS 16.0, *) {
                         ChartView()
+                    } else {
+                        ChartViewCompatibility()
                     }
                 }
