+++ AppLibrary/Extensions/UserDefaults.swift
@@ -47,7 +47,6 @@ private enum Keys: String {
     case sharedTransmitterFirmware = "glucosedirect--transmitter-firmware"
     case sharedTransmitterHardware = "glucosedirect--transmitter-hardware"
     case transmitter = "libre-direct.settings.transmitter"
-    case glucoseValues = "libre-direct.settings.glucose-values"
 }
 
 extension UserDefaults {
@@ -563,24 +562,6 @@ extension UserDefaults {
             } else {
                 removeObject(forKey: Keys.transmitter.rawValue)
             }
-        }
-    }
-    
-    var gen1GlucoseValues: [Gen1Glucose] {
-        get {
-            return getArray(forKey: Keys.glucoseValues.rawValue) ?? []
-        }
-        set {
-            setArray(newValue, forKey: Keys.glucoseValues.rawValue)
-        }
-    }
-
-    var gen2GlucoseValues: [Gen2Glucose] {
-        get {
-            return getArray(forKey: Keys.glucoseValues.rawValue) ?? []
-        }
-        set {
-            setArray(newValue, forKey: Keys.glucoseValues.rawValue)
         }
     }
 }
