+++ GlucoseDirect.xcconfig
@@ -6,6 +6,6 @@
 DEVELOPMENT_TEAM = ##TEAM_ID##
 PRODUCT_BUNDLE_IDENTIFIER = com.${DEVELOPMENT_TEAM}.libre-direct
 APP_GROUP_ID = group.com.$(DEVELOPMENT_TEAM).loopkit.LoopGroup
-APP_VERSION = 0.10.0
+APP_VERSION = 0.11.0
 
 #include? "GlucoseDirectOverride.xcconfig"
